# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tako Internal API
  version: 0.1.0
  description: Internal dashboard contract. Third-party compatibility is not guaranteed during v1.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session
        '401':
          $ref: '#/components/responses/Problem'
...
  /logs:
    get:
      operationId: listLogs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        '200': { description: Recent journal entries }
        '503': { $ref: '#/components/responses/Problem' }
...
  /services:
    get:
      operationId: listServices
      responses:
        '200': { description: systemd unit snapshot }
        '503': { $ref: '#/components/responses/Problem' }
...
  /terminal/ws:
    get:
      operationId: openTerminal
      parameters:
        - in: query
          name: columns
          schema: { type: integer, minimum: 1, maximum: 65535 }
        - in: query
          name: rows
          schema: { type: integer, minimum: 1, maximum: 65535 }
      responses:
        '101': { description: Binary PTY WebSocket }
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, maxLength: 256 }
        password: { type: string, maxLength: 4096 }
    Problem:
      type: object
      required: [title, status, code]
      properties:
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
  responses:
    Problem:
      description: Structured API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'